split heavier methods from Module.h to ModuleMethods.h
